+++ CHANGELOG.md
@@ -5,6 +5,9 @@ HTML Template Changelog
 -----
 
 *   BC break: Add method `\VovanVE\HtmlTemplate\compile\CompilerInterface::getMeta()`.
+*   BC break: Method `\VovanVE\HtmlTemplate\EngineInterface::runTemplate()` now receive
+    optional `\VovanVE\HtmlTemplate\runtime\RuntimeHelperInterface` in 2nd argument
+    instead of just `array`.
 *   Deprecated: constant `\VovanVE\HtmlTemplate\Engine::VERSION`.
 
 0.0.1